Move translation features into own module
