+++ src/datasets/features/__init__.py
@@ -1,1 +1,2 @@
 from .features import *
+from .translation import *
